requirements.txt
-e .
pandas
pyyaml
streamlit
